refactor(workflow): extract runtime worker tests
